requirements.txt
python-dotenv==1.0.1
google-cloud-storage==2.18.2
langchain==0.3.7
langchain-community==0.3.5
langchain-chroma==0.1.4
fastembed==0.4.1
chromadb==0.5.18
openai==1.54.3
Flask==3.0.3
Flask-Cors==5.0.0